Consistency update for fastjet-finder
Align the arguments for fastjet-finder to be the same as those for the
Julia test scripts, e.g., -m or --nsamples for the number of trial
repeats.

Ensure that algorithm and power values from benchmark.jl are correctly
passed to fastjet-finder. It seems the algorithm setting was actually
being _ignored_, meaning only pp-type algorithms actually got run.